rev-parse: document that most of the time you need to be in a repository
It may be confusing to new users that "git rev-parse
--is-inside-git-dir" gives a fatal error instead of showing "false"
when run outside a repository or a working tree controlled by a
repository.  Instead of saying that they should expect ternary,
i.e. an error, true, or false, tell them upfront that most operations
require you to be in a repository or a working tree of a repository.

Signed-off-by: Junio C Hamano <gitster@pobox.com>